+++ sdk/communication/azure-communication-administration/azure/communication/administration/aio/_phone_number_administration_client_async.py
@@ -38,7 +38,8 @@ from .._phonenumber._generated.models import (
     UpdatePhoneNumberCapabilitiesResponse
 )
 
-from .._shared.utils import parse_connection_str, get_authentication_policy
+from .._shared.utils import parse_connection_str
+from .._shared.policy import HMACCredentialsPolicy
 
 class PhoneNumberAdministrationClient(object):
     """Azure Communication Services Phone Number Management client.
@@ -69,7 +70,7 @@ class PhoneNumberAdministrationClient(object):
         self._endpoint = endpoint
         self._phone_number_administration_client = PhoneNumberAdministrationClientGen(
             self._endpoint,
-            authentication_policy=get_authentication_policy(endpoint, credential, is_async=True),
+            authentication_policy=HMACCredentialsPolicy(endpoint, credential),
             sdk_moniker=SDK_MONIKER,
             **kwargs)