+++ sdk/communication/azure-communication-administration/CHANGELOG.md
@@ -24,10 +24,9 @@ Caller must provide one of the following:
 - Added `MicrosoftTeamsUserIdentifier`
 
 ##### `IdentityClient`
-- Added support for Azure Active Directory authentication
+- Added support for Azure Active Directory authentication for the Identity client
 
-#### `PhoneNumberAdministrationClient`
-- Added support for Azure Active Directory authentication
+
 ## 1.0.0b3 (2020-11-16)
 
 ### Breaking Changes